move array_compatibility to misc
