+++ scipy/misc/array_compatibility.py
